# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Price Pulse API",
    "description": "Retail product price monitoring platform.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/products/{product_id}/prices": {
      "get": {
        "tags": [
          "prices"
        ],
        "summary": "Paginated price history for a product",
        "operationId": "list_prices_api_v1_products__product_id__prices_get",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Product Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Max items per page (\u2264 100)",
              "default": 20,
              "title": "Limit"
            },
            "description": "Max items per page (\u2264 100)"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip"
          },
          {
            "name": "from_dt",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO 8601 lower bound (inclusive)",
              "title": "From Dt"
            },
            "description": "ISO 8601 lower bound (inclusive)"
          },
          {
            "name": "to_dt",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO 8601 upper bound (inclusive)",
              "title": "To Dt"
            },
            "description": "ISO 8601 upper bound (inclusive)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_PriceRecordRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{product_id}/scrape": {
      "post": {
        "tags": [
          "prices"
        ],
        "summary": "Trigger an on-demand price scrape for a product",
        "operationId": "trigger_scrape_api_v1_products__product_id__scrape_post",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScrapeJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alerts": {
      "post": {
        "tags": [
          "alerts"
        ],
        "summary": "Create a price alert for a product",
        "operationId": "create_alert_api_v1_alerts_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "alerts"
        ],
        "summary": "List price alerts",
        "operationId": "list_alerts_api_v1_alerts_get",
        "parameters": [
          {
            "name": "product_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter alerts by product",
              "title": "Product Id"
            },
            "description": "Filter alerts by product"
          },
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by active status",
              "title": "Is Active"
            },
            "description": "Filter by active status"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Max items per page (\u2264 100)",
              "default": 20,
              "title": "Limit"
            },
            "description": "Max items per page (\u2264 100)"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_AlertRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alerts/{alert_id}": {
      "get": {
        "tags": [
          "alerts"
        ],
        "summary": "Retrieve a single alert",
        "operationId": "get_alert_api_v1_alerts__alert_id__get",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Alert Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "alerts"
        ],
        "summary": "Partially update an alert (product_id is immutable)",
        "operationId": "update_alert_api_v1_alerts__alert_id__patch",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Alert Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "alerts"
        ],
        "summary": "Delete a price alert",
        "operationId": "delete_alert_api_v1_alerts__alert_id__delete",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Alert Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AlertCreate": {
        "properties": {
          "product_id": {
            "type": "integer",
            "title": "Product Id"
          },
          "threshold_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Threshold Price"
          },
          "direction": {
            "$ref": "#/components/schemas/AlertDirection"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "channel": {
            "$ref": "#/components/schemas/NotificationChannel",
            "default": "email"
          },
          "webhook_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhook Url"
          },
          "whatsapp_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Whatsapp Number"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "threshold_price",
          "direction"
        ],
        "title": "AlertCreate"
      },
      "AlertDirection": {
        "type": "string",
        "enum": [
          "above",
          "below"
        ],
        "title": "AlertDirection"
      },
      "AlertRead": {
        "properties": {
          "product_id": {
            "type": "integer",
            "title": "Product Id"
          },
          "threshold_price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Threshold Price"
          },
          "direction": {
            "$ref": "#/components/schemas/AlertDirection"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "channel": {
            "$ref": "#/components/schemas/NotificationChannel",
            "default": "email"
          },
          "webhook_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhook Url"
          },
          "whatsapp_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Whatsapp Number"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "notified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notified At"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "threshold_price",
          "direction",
          "id"
        ],
        "title": "AlertRead"
      },
      "AlertUpdate": {
        "properties": {
          "threshold_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threshold Price"
          },
          "direction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AlertDirection"
              },
              {
                "type": "null"
              }
            ]
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "channel": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NotificationChannel"
              },
              {
                "type": "null"
              }
            ]
          },
          "webhook_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhook Url"
          },
          "whatsapp_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Whatsapp Number"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AlertUpdate",
        "description": "Partial-update schema for PriceAlert.\n\n``product_id`` is intentionally absent \u2014 alert ownership is immutable after\ncreation.  Passing ``product_id`` in the request body returns 422."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "NotificationChannel": {
        "type": "string",
        "enum": [
          "email",
          "webhook",
          "whatsapp"
        ],
        "title": "NotificationChannel"
      },
      "PaginatedResponse_AlertRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AlertRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "PaginatedResponse[AlertRead]"
      },
      "PaginatedResponse_PriceRecordRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PriceRecordRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "PaginatedResponse[PriceRecordRead]"
      },
...
      "PriceRecordRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "product_id": {
            "type": "integer",
            "title": "Product Id"
          },
          "price": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "captured_at": {
            "type": "string",
            "format": "date-time",
            "title": "Captured At"
          },
          "raw_html_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Html Hash"
          },
          "extraction_status": {
            "type": "string",
            "title": "Extraction Status",
            "default": "ok"
          }
        },
        "type": "object",
        "required": [
          "id",
          "product_id",
          "captured_at"
        ],
        "title": "PriceRecordRead",
        "description": "Read schema for PriceRecord.\n\n``price`` and ``currency`` are nullable \u2014 failed scrape attempts are stored\nwith ``price=NULL`` and ``currency=NULL`` (see Item 4 migration).\n``extraction_status`` reflects the outcome of the scrape attempt."
      },
...
      "ProductRead": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "source_type": {
            "$ref": "#/components/schemas/SourceType"
          },
          "css_selector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Css Selector"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "url",
          "source_type",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "ProductRead"
      },
...
      "ScrapeJobResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "status": {
            "type": "string",
            "const": "queued",
            "title": "Status"
          },
          "product": {
            "$ref": "#/components/schemas/ProductRead"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "status",
          "product"
        ],
        "title": "ScrapeJobResponse",
        "description": "Response body for POST /products/{id}/scrape (202 Accepted).\n\n``task_id`` is the Celery task ID; the caller can use it to poll for results.\n``product`` reflects the current product state so the UI can render\nimmediately without a second GET request."
      },
      "SourceType": {
        "type": "string",
        "enum": [
          "generic",
          "amazon",
          "ebay",
          "currys"
        ],
        "title": "SourceType"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}